refactor: Structure Rust code in different modules
