+++ services/webapp-rust/src/core/constants.rs
@@ -1,0 +1,9 @@
+pub const PORT_NUM: u16 = 3000;
+pub const APP_VERSION: &'static str = "v0.1.0";
+pub const ENTRY_POINT_DIR_NAME: &'static str = "programm"; // arbitrary name given by vmassimi
+
+pub const ARCHIVES_ROOT_DIR: &'static str = "/app/data/archives";
+pub const ARCHIVES_TMP_DIR: &'static str = "/app/data/archives/tmp";
+pub const VERSIONS_PATH: &'static str = "/app/data/versions.json";
+
+pub const ZFILL_PADDING: usize = 3;
